Enabling viewport; Corrections and how render a Framebuffer; Adding framebuffer to ImGui viewport.

+++ src/modules/sparkengine/Source/Private/Spark/Metal/MetalFramebuffer.h
@@ -10,8 +10,8 @@ namespace Spark
     private:
         FramebufferProperties m_Properties;
         uint32 m_RendererID = 0;
-        uint32 m_ColorAttachmentRendererID = 0;
-        uint32 m_DepthAttachmentRendererID = 0;
+        uint64 m_ColorAttachmentRendererID = 0;
+        uint64 m_DepthAttachmentRendererID = 0;
         
         MTLRenderPassDescriptor* m_RenderPassDescriptor = nil;
         id<MTLTexture> m_ColorTexture;
@@ -27,8 +27,8 @@ namespace Spark
         virtual void Unbind() override;
         virtual void Invalidate() override;
         virtual uint32 GetRendererID() override { return m_RendererID; }
-        virtual uint32 GetColorAttachmentRendererID() override { return m_ColorAttachmentRendererID; }
-        virtual uint32 GetDepthAttachmentRendererID() override { return m_DepthAttachmentRendererID; }
+        virtual uint64 GetColorAttachmentRendererID() override { return m_ColorAttachmentRendererID; }
+        virtual uint64 GetDepthAttachmentRendererID() override { return m_DepthAttachmentRendererID; }
         virtual const FramebufferProperties &GetProperties() const override {return m_Properties; }
         virtual void SetProperties(const FramebufferProperties& Properties) override;
         virtual void Resize(uint32 width, uint32 height) override;
