+++ src/modules/sparkengine/Source/Private/Spark/OpenGL/OpenGLFramebuffer.h
@@ -20,8 +20,8 @@ namespace Spark
         virtual void Unbind() override;
         virtual void Invalidate() override;
         virtual uint32 GetRendererID() override { return m_RendererID; }
-        virtual uint32 GetColorAttachmentRendererID() override { return m_ColorAttachmentRendererID; }
-        virtual uint32 GetDepthAttachmentRendererID() override { return m_DepthAttachmentRendererID; }
+        virtual uint64 GetColorAttachmentRendererID() override { return m_ColorAttachmentRendererID; }
+        virtual uint64 GetDepthAttachmentRendererID() override { return m_DepthAttachmentRendererID; }
         virtual const FramebufferProperties &GetProperties() const override {return m_Properties; }
         virtual void SetProperties(const FramebufferProperties& Properties) override;
         virtual void Resize(uint32 width, uint32 height) override;
